Move to named export and switch to using styled API
