--- src/components/Dashboard/ThisWeek/PartnerCare/index.tsx
@@ -1,3 +1,0 @@
-import PartnerCare from './PartnerCare';
-
-export default PartnerCare;
